gallery-server/requirements.txt
autopep8==2.0.0
click==8.1.3
Flask==2.2.2
flask-oidc==1.4.0
httplib2==0.21.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.2
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
rsa==4.9
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.2
Pillow==9.4.0
Flask-Cors==3.0.10
